Unit tests for kernel defaults
