+++ dev/com.ibm.ws.kernel.boot_test/test/test data/default-files/default/lib/platform/defaults
@@ -1,0 +1,6 @@
+# Use the following properties (in the System or bootstrap.properties) to
+# override the values:
+#   websphere.kernel
+#   websphere.log.provider
+WebSphere-DefaultKernel: kernelCore-1.0
+WebSphere-DefaultLogProvider: defaultLogging-1.0
