+++ dev/com.ibm.ws.kernel.boot_test/test/test data/default-files/extension/lib/platform/defaults
@@ -1,0 +1,3 @@
+# These defaults are used by the test bucket only.
+WebSphere-DefaultKernel: kernelCoreTest-1.0
+WebSphere-DefaultLogProvider: defaultLoggingTest-1.0
