+++ dev/com.ibm.ws.kernel.boot_test/test/test data/default-files/missing/lib/platform/defaults
@@ -1,0 +1,1 @@
+# These defaults are used by the test bucket only and are intentionally blank.
